+++ src/game/constants.rs
@@ -4,9 +4,9 @@ pub const ROUND_LIMIT: usize = 30;
     
 // Ship Properties
 pub const START_COAL: usize = 6;
-pub const MIN_SPEED: usize = 1;
-pub const MAX_SPEED: usize = 6;
-pub const FREE_ACC: usize = 1;
+pub const MIN_SPEED: i32 = 1;
+pub const MAX_SPEED: i32 = 6;
+pub const FREE_ACC: i32 = 1;
 
 // Points
 pub const FINISH_POINTS: usize = 6;
